backend/requirements.txt
Django==5.0.7
psycopg==3.2.1
django-filter==24.2
djangorestframework==3.15.2
requests==2.32.3
openmeteo-requests==1.2.0
numpy==2.0.0
pandas==2.2.2
requests-cache==1.2.1
retry-requests==2.0.0
